Implemented Prim's algorithm to find minimum spanning tree for the given graph

+++ dijkstra_graph_path.cpp
@@ -45,7 +45,7 @@ public:
     // using a kind of A* algorithm
     {
         std::priority_queue<PathWithPriority> pq;
-        std::set<int> visited;
+        std::set<size_t> visited;
 
         pq.push(x);
